feat: move login_required wrapper outside (#1281)
