+++ api/controllers/console/datasets/hit_testing.py
@@ -1,7 +1,7 @@
 import logging
 
 from flask_login import current_user
-from core.login.login import login_required
+from libs.login import login_required
 from flask_restful import Resource, reqparse, marshal
 from werkzeug.exceptions import InternalServerError, NotFound, Forbidden
